+++ CHANGELOG.md
@@ -15,12 +15,14 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 - Added IE specific styling to `globals.css` for the `/notsupported` page
 - Added IE detection to `/404` and `/500` pages as they don't auto redirect using the next.js redirects function
 - Added language detection to redirect user to the page in prefered language
+- Added a signup info page prior to the signup page
 
 # Changed
 
 - Lock scroll when feedback module is active
 - Deactivate auto focus when feedback module is active
 - Updated introduction paragraph on the project page
+- Removed all the initial text on signup page
 
 ## Fixed